Refactor `HoldConfirmation`
