+++ .babelrcExp
@@ -62,7 +62,7 @@
           "@FieldsetDate": "./src/app/components/Filters/FieldsetDate.jsx",
           "@FieldsetList": "./src/app/components/Filters/FieldsetList.jsx",
           "@SelectedFilters": "./src/app/components/Filters/SelectedFilters.jsx",
-          "@HoldConfirmation": "./src/app/components/HoldConfirmation/HoldConfirmation.jsx",
+          "@HoldConfirmation": "./src/app/pages/HoldConfirmation.jsx",
           "@HoldRequest": "./src/app/pages/HoldRequest.jsx",
           "@Home": "./src/app/pages/Home.jsx",
           "@ItemsContainer": "./src/app/components/Item/ItemsContainer.jsx",
@@ -146,7 +146,6 @@
           "@Alt": "./test/unit/Alt.test.js",
           "@unitApplication": "./test/unit/Application.test.js",
           "@unitBibDetails": "./test/unit/BibDetails.test.js",
-          "@unitBreadcrumbs": "./test/unit/Breadcrumbs.test.js",
           "@unitFieldsetDate": "./test/unit/FieldsetDate.test.js",
           "@unitFieldsetList": "./test/unit/FieldsetList.test.js",
           "@unitFilterPopup": "./test/unit/FilterPopup.test.js",
